Update cost for hash and cached joins
The old code didn't correctly add TIME_FOR_COMPARE to rows that are
part of the scan that will be compared with the attached where clause.

Now the cost calculation for hash join and full join cache join are
identical except for HASH_FANOUT (10%)

Best_cost now includes the cost of comparing all WHERE clauses and also
cost of joining with previous row combinations.