MDEV-26668 CHECK TABLE fails to report schema mismatch
check_index_consistency(): when no indexes are defined explicitly for a table,
check clustered index fields against table definion

innobase_match_index_columns(): the change is not a refactoring only.
Notice a while loop which was removed. Once it encounters a DATA_SYS
field it skips all the left fields! This is fixed now.